add support for overlayfs

+++ common/build.sh
@@ -254,11 +254,10 @@ function copy_and_verify(){
 function prepare_image_for_friendlyelec_eflasher(){
     local OS_DIR=$1
     local ROOTFS=$2
-    local IMGSIZE=$3
     if [ ! -d ${SDFUSE_DIR}/${OS_DIR} ]; then
         mkdir ${SDFUSE_DIR}/${OS_DIR}
     fi
-    rm -f ${SDFUSE_DIR}/${OS_DIR}/*
+    rm -rf ${SDFUSE_DIR}/${OS_DIR}/*
 
     copy_and_verify $TOP_DIR/u-boot/rk3399_loader_v1.12.109.bin ${SDFUSE_DIR}/${OS_DIR}/MiniLoaderAll.bin "error: please build uboot first."
     copy_and_verify $TOP_DIR/u-boot/uboot.img ${SDFUSE_DIR}/${OS_DIR} "error: please build uboot first."
@@ -299,7 +298,6 @@ function prepare_image_for_friendlyelec_eflasher(){
 
     cat << 'EOL' > $FA_TMP_DIR/rootfs/etc/fstab
     # <file system>                 <mount pt>              <type>          <options>               <dump>  <pass>
-    /dev/root                       /                       ext2            rw,noauto               0       1
     proc                            /proc                   proc            defaults                0       0
     devpts                          /dev/pts                devpts          defaults,gid=5,mode=620 0       0
     tmpfs                           /dev/shm                tmpfs           mode=0777               0       0
@@ -337,14 +335,38 @@ EOL
     # remove /dev/console
     rm -f $FA_TMP_DIR/rootfs/dev/console
 
-    (cd $FA_TMP_DIR/ && ${SDFUSE_DIR}/tools/make_ext4fs -s -l ${IMGSIZE} -a root -L rootfs rootfs.img rootfs)
+
+    MKFS=${SDFUSE_DIR}/tools/make_ext4fs
+    ROOT=${FA_TMP_DIR}/rootfs
+
+    # calc image size
+    ROOTFS_SIZE=`du -s -B 1 ${ROOT} | cut -f1`
+    MAX_IMG_SIZE=7100000000
+    TMPFILE=`tempfile`
+    ${MKFS} -s -l ${MAX_IMG_SIZE} -a root -L rootfs /dev/null ${ROOT} > ${TMPFILE}
+    IMG_SIZE=`cat ${TMPFILE} | grep "Suggest size:" | cut -f2 -d ':' | awk '{gsub(/^\s+|\s+$/, "");print}'`
+    cat $TMPFILE
+    rm -f ${TMPFILE}
+
+    if [ ${ROOTFS_SIZE} -gt ${IMG_SIZE} ]; then
+            pt_error "IMG_SIZE less than ROOTFS_SIZE, why?"
+            exit 1
+    fi
+
+    # make rootfs.img
+    ${MKFS} -s -l ${IMG_SIZE} -a root -L rootfs $FA_TMP_DIR/rootfs.img ${ROOT}
     copy_and_verify $FA_TMP_DIR/rootfs.img ${SDFUSE_DIR}/${OS_DIR}/rootfs.img " "
     rm -rf $FA_TMP_DIR
+
+    # fill size to parameter.txt and partmap.txt
+    ROOTFS_PARTITION_SIZE=`printf "0x%X" $(($IMG_SIZE/512))`
+    ROOTFS_PARTITION_ADDR=0x00030000
+    USERDATA_PARTITION_ADDR=`printf "0x%X" $((${ROOTFS_PARTITION_ADDR}+${ROOTFS_PARTITION_SIZE}))`
 
     ###############################
     ######## parameter.txt ########
     ###############################
-cat << 'EOL' > ${SDFUSE_DIR}/${OS_DIR}/parameter.txt
+cat > ${SDFUSE_DIR}/${OS_DIR}/parameter.txt << EOL
 FIRMWARE_VER: 6.0.1
 MACHINE_MODEL: RK3399
 MACHINE_ID: 007
@@ -358,9 +380,61 @@ PWR_HLD: 0,0,A,0,1
 #FDT_NAME: rk-kernel.dtb
 #RECOVER_KEY: 1,1,0,20,0
 #in section; per section 512(0x200) bytes
-CMDLINE: root=/dev/mmcblk1p7 rw rootfstype=ext4 mtdparts=rk29xxnand:0x00002000@0x00002000(uboot),0x00002000@0x00004000(trust),0x00002000@0x00006000(misc),0x00006000@0x00008000(resource),0x00010000@0x0000e000(kernel),0x00010000@0x0001e000(boot),-@0x00030000(rootfs)
+CMDLINE: root=/dev/mmcblk1p7 rw rootfstype=ext4 data=/dev/mmcblk1p8 consoleblank=0 mtdparts=rk29xxnand:0x00002000@0x00002000(uboot),0x00002000@0x00004000(trust),0x00002000@0x00006000(misc),0x00006000@0x00008000(resource),0x00010000@0x0000e000(kernel),0x00010000@0x0001e000(boot),${ROOTFS_PARTITION_SIZE}@0x00030000(rootfs),-@${USERDATA_PARTITION_ADDR}(userdata)
 
 EOL
+
+    mkdir -p ${SDFUSE_DIR}/${OS_DIR}/sd-boot
+    ##########################################################
+    ######## param4sd.txt (parameter.txt for sd-boot) ########
+    ##########################################################
+cat > ${SDFUSE_DIR}/${OS_DIR}/sd-boot/param4sd.txt << EOL
+FIRMWARE_VER: 6.0.1
+MACHINE_MODEL: RK3399
+MACHINE_ID: 007
+MANUFACTURER: RK3399
+MAGIC: 0x5041524B
+ATAG: 0x00200800
+MACHINE: 3399
+CHECK_MASK: 0x80
+PWR_HLD: 0,0,A,0,1
+#KERNEL_IMG: 0x00280000
+#FDT_NAME: rk-kernel.dtb
+#RECOVER_KEY: 1,1,0,20,0
+#in section; per section 512(0x200) bytes
+CMDLINE: root=/dev/mmcblk0p1 rw rootfstype=ext4 data=/dev/mmcblk0p2 consoleblank=0 mtdparts=rk29xxnand:0x00002000@0x00002000(uboot),0x00002000@0x00004000(trust),0x00002000@0x00006000(misc),0x00006000@0x00008000(resource),0x00010000@0x0000e000(kernel),0x00010000@0x0001e000(boot),${ROOTFS_PARTITION_SIZE}@0x00030000(rootfs),-@${USERDATA_PARTITION_ADDR}(userdata)
+
+EOL
+
+    ROOTFS_PARTITION_SIZE=`printf "0x%X" ${IMG_SIZE}`
+    ROOTFS_PARTITION_ADDR=0x6000000
+    USERDATA_PARTITION_ADDR=`printf "0x%X" $((${ROOTFS_PARTITION_ADDR}+${ROOTFS_PARTITION_SIZE}))`
+
+    #########################################
+    ######## partmap.txt for sd-boot ########
+    #########################################
+cat > ${SDFUSE_DIR}/${OS_DIR}/sd-boot/partmap.txt << EOL
+flash=mmc,1:loader:idb:0x8000,0x280000:idbloader.img;
+flash=mmc,1:env:env:0x3F8000,0x8000;
+flash=mmc,1:parm:parm:0x400000,0x0400000:param4sd.txt;
+flash=mmc,1:uboot:raw:0x800000,0x0400000:uboot.img;
+flash=mmc,1:trust:raw:0xC00000,0x0400000:trust.img;
+flash=mmc,1:misc:raw:0x1000000,0x0400000;
+flash=mmc,1:resc:raw:0x1400000,0x0C00000:resource.img;
+flash=mmc,1:kern:raw:0x2000000,0x2000000:kernel.img;
+flash=mmc,1:boot:raw:0x4000000,0x2000000:boot.img;
+flash=mmc,1:rootfs:ext4:0x6000000,${ROOTFS_PARTITION_SIZE}:rootfs.img;
+flash=mmc,1:userdata:ext4:${USERDATA_PARTITION_ADDR},0x0:userdata.img;
+
+EOL
+
+    # gen empty userdata.img
+    if [ ! -f ${SDFUSE_DIR}/${OS_DIR}/userdata.img ]; then
+        USERDATA_SIZE=104857600
+        TMPDIR=`mktemp -d`
+        ${MKFS} -s -l ${USERDATA_SIZE} -a root -L userdata ${SDFUSE_DIR}/${OS_DIR}/userdata.img ${TMPDIR}
+        rm -rf ${TMPDIR}
+    fi
 }
 
 function clean_device_files()
@@ -377,18 +451,15 @@ function clean_device_files()
 function build_sdimg(){
     local IMG=
     local ROOTFS=
-    local IMGSIZE=
     if [ -z ${OS_NAME} ]; then
         OS_NAME=buildroot
     fi
     if [ x${OS_NAME} = xbuildroot ]; then
         IMG=rk3399-sd-buildroot-linux-4.4-arm64-$(date +%Y%m%d).img
         ROOTFS=${TOP_DIR}/buildroot/output/rockchip_rk3399/images/rootfs.ext2
-        IMGSIZE=1073741824
     elif [ x${OS_NAME} = xdebian ]; then
         IMG=rk3399-sd-debian9-4.4-armhf-$(date +%Y%m%d).img
         ROOTFS=${TOP_DIR}/rootfs/binary
-        IMGSIZE=3670016000
         if [ ! -d ${TOP_DIR}/rootfs/binary ]; then
             echo "error: please build debian first."
             exit 1
@@ -397,7 +468,7 @@ function build_sdimg(){
         echo "Unknow OS: ${OS_NAME}"
         exit 1
     fi
-    prepare_image_for_friendlyelec_eflasher ${OS_NAME} ${ROOTFS} ${IMGSIZE} && (cd ${SDFUSE_DIR} && ./mk-sd-image.sh ${OS_NAME}) && {
+    prepare_image_for_friendlyelec_eflasher ${OS_NAME} ${ROOTFS} && (cd ${SDFUSE_DIR} && ./mk-sd-image.sh ${OS_NAME}) && {
         echo "-----------------------------------------"
         echo "Run the following for sdcard install:"
         echo "    dd if=out/${IMG} of=/dev/sdX bs=1M"
@@ -408,23 +479,20 @@ function build_sdimg(){
 function build_emmcimg() {
     local IMG=
     local ROOTFS=
-    local IMGSIZE=
     if [ -z ${OS_NAME} ]; then
         OS_NAME=buildroot
     fi
     if [ x${OS_NAME} = xbuildroot ]; then
         IMG=rk3399-eflasher-buildroot-linux-4.4-arm64-$(date +%Y%m%d).img
         ROOTFS=${TOP_DIR}/buildroot/output/rockchip_rk3399/images/rootfs.ext2
-        IMGSIZE=1073741824
     elif [ x${OS_NAME} = xdebian ]; then
         IMG=rk3399-eflasher-debian9-4.4-armhf-$(date +%Y%m%d).img
         ROOTFS=${TOP_DIR}/rootfs/binary
-        IMGSIZE=3670016000
     else
         echo "Unknow OS: ${OS_NAME}"
         exit 1
     fi
-    prepare_image_for_friendlyelec_eflasher ${OS_NAME} ${ROOTFS} ${IMGSIZE} && (cd ${SDFUSE_DIR} && ./mk-emmc-image.sh ${OS_NAME}) && {
+    prepare_image_for_friendlyelec_eflasher ${OS_NAME} ${ROOTFS} && (cd ${SDFUSE_DIR} && ./mk-emmc-image.sh ${OS_NAME}) && {
         echo "-----------------------------------------"
         echo "Run the following for sdcard install:"
         echo "    dd if=out/${IMG} of=/dev/sdX bs=1M"
